requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.4.0
sqlparse==0.5.0